backend/msg-handler/requirements.txt
Flask==3.0.3 
Flask-Cors==5.0.0
paho-mqtt==2.1.0